Remove z-index: 3
This was causing a bug in My Courses where a course tile's menu could end up underneath a different course's menu. This z-index was added in the first place to fix a different problem there, but another change allows us to avoid the circumstance altogether, removing the need for this z-index.